REFACTOR to distribute different components on different files
to keep order :)
