+++ src/pages/Admin/Results/ResumeElection.jsx
@@ -1,0 +1,11 @@
+import ResumeTable from "../ElectionResume/components/ResumeTable";
+import CardTitle from "./components/CardTitle";
+
+export default function ResumeElection() {
+    return (
+        <>
+            <CardTitle title="Resumen elección" />
+            <ResumeTable className="pt-4" />
+        </>
+    );
+}
