+++ src/pages/Admin/Results/components/CardTitle.jsx
@@ -1,0 +1,7 @@
+export default function CardTitle({ title }) {
+    return (
+        <div className="d-flex py-2">
+            <h1 className="title is-size-3">{title}</h1>
+        </div>
+    );
+}
